fix for condition 2

+++ Estimation/CellsDataContainer.cpp
@@ -124,7 +124,7 @@ namespace Estimation
 		if (cells_genes_counts.size() > 0)
 		{
 			ss << "top CBs:\n";
-			size_t low_border = cells_genes_counts.size() - min(cells_genes_counts.size(), this->_top_print_size - 1);
+			long low_border = cells_genes_counts.size() - min(cells_genes_counts.size(), this->_top_print_size - 1);
 			for (long i = cells_genes_counts.size() - 1; i >= low_border; --i)
 			{
 				ss << cells_genes_counts[i].value << "\t" << this->_cell_barcodes[cells_genes_counts[i].index] << "\n";
